@@ -1,16 +1,18 @@
-import {combineReducers, createStore} from "redux";
-import {tasksReducer} from "./tasks-reducer";
-import {todolistsReducer} from "./todolists-reducer";
+import { tasksReducer } from './tasks-reducer';
+import { todolistsReducer } from './todolists-reducer';
+import { combineReducers, createStore } from 'redux';
 
+// объединяя reducer-ы с помощью combineReducers,
+// мы задаём структуру нашего единственного объекта-состояния
 const rootReducer = combineReducers({
-    todolists: todolistsReducer,
-    tasks: tasksReducer
+    tasks: tasksReducer,
+    todolists: todolistsReducer
 })
-
-export type AppRootState = ReturnType<typeof rootReducer>
-
+// непосредственно создаём store
 export const store = createStore(rootReducer);
+// определить автоматически тип всего объекта состояния
+export type AppRootStateType = ReturnType<typeof rootReducer>
 
+// а это, чтобы можно было в консоли браузера обращаться к store в любой момент
 // @ts-ignore
-window.store = store
-
+window.store = store;
