@@ -1,40 +1,57 @@
-import {v1} from "uuid";
-import {FilterValuesType, TodolistType} from "../AppWithRedux";
+import { v1 } from 'uuid';
+import { FilterValuesType, TodolistType } from '../App';
 
-export type ActionsType =
-    ReturnType<typeof removeTodolistAC> |
-    ReturnType<typeof addTodolistAC> |
-    ReturnType<typeof changeTodolistTitleAC> |
-    ReturnType<typeof changeTodolistFilterAC>
+export type RemoveTodolistActionType = {
+    type: 'REMOVE-TODOLIST',
+    id: string
+}
+export type AddTodolistActionType = {
+    type: 'ADD-TODOLIST',
+    title: string
+    todolistId: string
+}
+export type ChangeTodolistTitleActionType = {
+    type: 'CHANGE-TODOLIST-TITLE',
+    id: string
+    title: string
+}
+export type ChangeTodolistFilterActionType = {
+    type: 'CHANGE-TODOLIST-FILTER',
+    id: string
+    filter: FilterValuesType
+}
+
+type ActionsType = RemoveTodolistActionType | AddTodolistActionType
+    | ChangeTodolistTitleActionType
+    | ChangeTodolistFilterActionType
 
 const initialState: Array<TodolistType> = []
 
 export const todolistsReducer = (state: Array<TodolistType> = initialState, action: ActionsType): Array<TodolistType> => {
     switch (action.type) {
         case 'REMOVE-TODOLIST': {
-            return state.filter(f => f.id !== action.id)
+            return state.filter(tl => tl.id !== action.id)
         }
-
         case 'ADD-TODOLIST': {
             return [{
                 id: action.todolistId,
                 title: action.title,
                 filter: 'all'
-            }, ...state,]
+            }, ...state]
         }
-
         case 'CHANGE-TODOLIST-TITLE': {
-            let todolist = state.find(f => f.id === action.id)
+            const todolist = state.find(tl => tl.id === action.id);
             if (todolist) {
-                todolist.title = action.title
+                // если нашёлся - изменим ему заголовок
+                todolist.title = action.title;
             }
             return [...state]
         }
-
         case 'CHANGE-TODOLIST-FILTER': {
-            let todolist = state.find(f => f.id === action.todolistId)
+            const todolist = state.find(tl => tl.id === action.id);
             if (todolist) {
-                todolist.filter = action.filter
+                // если нашёлся - изменим ему заголовок
+                todolist.filter = action.filter;
             }
             return [...state]
         }
@@ -43,16 +60,16 @@ export const todolistsReducer = (state: Array<TodolistType> = initialState, acti
     }
 }
 
-export const removeTodolistAC = (id: string) => ({type: 'REMOVE-TODOLIST', id}) as const
-export const addTodolistAC = (title: string) => ({type: 'ADD-TODOLIST', title, todolistId: v1()}) as const
-export const changeTodolistTitleAC = (id: string, title: string) => ({
-    type: 'CHANGE-TODOLIST-TITLE',
-    id,
-    title
-}) as const
-export const changeTodolistFilterAC = (todolistId: string, filter: FilterValuesType) => ({
-    type: 'CHANGE-TODOLIST-FILTER',
-    todolistId,
-    filter
-}) as const
+export const removeTodolistAC = (todolistId: string): RemoveTodolistActionType => {
+    return {type: 'REMOVE-TODOLIST', id: todolistId}
+}
+export const addTodolistAC = (title: string): AddTodolistActionType => {
+    return {type: 'ADD-TODOLIST', title: title, todolistId: v1()}
+}
+export const changeTodolistTitleAC = (id: string, title: string): ChangeTodolistTitleActionType => {
+    return {type: 'CHANGE-TODOLIST-TITLE', id: id, title: title}
+}
+export const changeTodolistFilterAC = (id: string, filter: FilterValuesType): ChangeTodolistFilterActionType => {
+    return {type: 'CHANGE-TODOLIST-FILTER', id: id, filter: filter}
+}
 
